Move main.go to cmd/, follow go convention about projects layout and command line apps ('go get' friendly)
